+++ piccolo_api/media/base.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import abc
 import asyncio
+import itertools
 import logging
 import pathlib
 import string
@@ -67,6 +68,8 @@ ALLOWED_CHARACTERS = (
     "-",
     "_",
     ".",
+    "(",
+    ")",
 )
 
 
@@ -200,6 +203,8 @@ class MediaStorage(metaclass=abc.ABCMeta):
         Stores the file in whichever storage you're using, and returns a key
         which uniquely identifes the file.
 
+        :param file_name:
+            The file name with which the file will be stored.
         :param file:
             The file to store.
         :param user:
@@ -259,7 +264,11 @@ class MediaStorage(metaclass=abc.ABCMeta):
         Returns the file key for each file we have in the database.
         """
         table = self.column._meta.table
-        return await table.select(self.column).output(as_list=True)
+        response = await table.select(self.column).output(as_list=True)
+        if isinstance(self.column, Array):
+            return [i for i in itertools.chain(*response)]
+        else:
+            return response
 
     async def get_unused_file_keys(self) -> t.List[str]:
         """