+++ piccolo_api/media/s3.py
@@ -11,6 +11,7 @@ from piccolo.apps.user.tables import BaseUser
 from piccolo.columns.column_types import Array, Text, Varchar
 
 from .base import ALLOWED_CHARACTERS, ALLOWED_EXTENSIONS, MediaStorage
+from .content_type import CONTENT_TYPE
 
 if t.TYPE_CHECKING:  # pragma: no cover
     from concurrent.futures._base import Executor
@@ -21,9 +22,11 @@ class S3MediaStorage(MediaStorage):
         self,
         column: t.Union[Text, Varchar, Array],
         bucket_name: str,
-        folder_name: str,
+        folder_name: t.Optional[str] = None,
         connection_kwargs: t.Dict[str, t.Any] = None,
+        sign_urls: bool = True,
         signed_url_expiry: int = 3600,
+        upload_metadata: t.Dict[str, t.Any] = None,
         executor: t.Optional[Executor] = None,
         allowed_extensions: t.Optional[t.Sequence[str]] = ALLOWED_EXTENSIONS,
         allowed_characters: t.Optional[t.Sequence[str]] = ALLOWED_CHARACTERS,
@@ -35,18 +38,18 @@ class S3MediaStorage(MediaStorage):
         besides from Amazon Web Services.
 
         :param column:
-            The Piccolo ``Column`` which the storage is for.
+            The Piccolo :class:`Column <piccolo.columns.base.Column>` which the
+            storage is for.
         :param bucket_name:
             Which S3 bucket the files are stored in.
-        :param folder:
+        :param folder_name:
             The files will be stored in this folder within the bucket. S3
             buckets don't really have folders, but if ``folder`` is
             ``'movie_screenshots'``, then we store the file at
             ``'movie_screenshots/my-file-abc-123.jpeg'``, to simulate it being
             in a folder.
         :param connection_kwargs:
-            These kwargs are passed directly to ``boto3``. Learn more about
-            `available options <https://boto3.amazonaws.com/v1/documentation/api/latest/reference/core/session.html#boto3.session.Session.client>`_.
+            These kwargs are passed directly to the boto3 :meth:`client <boto3.session.Session.client>`.
             For example::
 
                 S3MediaStorage(
@@ -58,10 +61,52 @@ class S3MediaStorage(MediaStorage):
                         'region_name': 'uk'
                     }
                 )
-
+        :param sign_urls:
+            Whether to sign the URLs - by default this is ``True``, as it's
+            highly recommended that your store your files in a private bucket.
         :param signed_url_expiry:
             Files are accessed via signed URLs, which are only valid for this
             number of seconds.
+        :param upload_metadata:
+            You can provide additional metadata to the uploaded files. To
+            see all available options see :class:`S3Transfer.ALLOWED_UPLOAD_ARGS <boto3.s3.transfer.S3Transfer>`.
+            Below we show examples of common use cases.
+
+            To set the ACL::
+
+                S3MediaStorage(
+                    ...,
+                    upload_metadata={'ACL': 'my_acl'}
+                )
+
+            To set the content disposition (how the file behaves when opened -
+            is it downloaded, or shown in the browser)::
+
+                S3MediaStorage(
+                    ...,
+                    # Shows the file within the browser:
+                    upload_metadata={'ContentDisposition': 'inline'}
+                )
+
+            To attach `user defined metadata <https://docs.aws.amazon.com/AmazonS3/latest/userguide/UsingMetadata.html>`_
+            to the file::
+
+                S3MediaStorage(
+                    ...,
+                    upload_metadata={'Metadata': {'myfield': 'abc123'}}
+                )
+
+            To specify how long browsers should cache the file for::
+
+                S3MediaStorage(
+                    ...,
+                    # Cache the file for 24 hours:
+                    upload_metadata={'CacheControl': 'max-age=86400'}
+                )
+
+            Note: We automatically add the ``ContentType`` field based on the
+            file type.
+
         :param executor:
             An executor, which file save operations are run in, to avoid
             blocking the event loop. If not specified, we use a sensibly
@@ -85,8 +130,10 @@ class S3MediaStorage(MediaStorage):
             self.boto3 = boto3
 
         self.bucket_name = bucket_name
+        self.upload_metadata = upload_metadata
         self.folder_name = folder_name
         self.connection_kwargs = connection_kwargs
+        self.sign_urls = sign_urls
         self.signed_url_expiry = signed_url_expiry
         self.executor = executor or ThreadPoolExecutor(max_workers=10)
 
@@ -96,15 +143,13 @@ class S3MediaStorage(MediaStorage):
             allowed_characters=allowed_characters,
         )
 
-    def get_client(self):  # pragma: no cover
+    def get_client(self, config=None):  # pragma: no cover
         """
-        Returns an S3 clent.
+        Returns an S3 client.
         """
         session = self.boto3.session.Session()
-        client = session.client(
-            "s3",
-            **self.connection_kwargs,
-        )
+        extra_kwargs = {"config": config} if config else {}
+        client = session.client("s3", **self.connection_kwargs, **extra_kwargs)
         return client
 
     async def store_file(
@@ -120,6 +165,13 @@ class S3MediaStorage(MediaStorage):
 
         return file_key
 
+    def _prepend_folder_name(self, file_key: str) -> str:
+        folder_name = self.folder_name
+        if folder_name:
+            return str(pathlib.Path(folder_name, file_key))
+        else:
+            return file_key
+
     def store_file_sync(
         self, file_name: str, file: t.IO, user: t.Optional[BaseUser] = None
     ) -> str:
@@ -127,13 +179,18 @@ class S3MediaStorage(MediaStorage):
         A sync wrapper around :meth:`store_file`.
         """
         file_key = self.generate_file_key(file_name=file_name, user=user)
-
+        extension = file_key.rsplit(".", 1)[-1]
         client = self.get_client()
+        upload_metadata: t.Dict[str, t.Any] = self.upload_metadata or {}
+
+        if extension in CONTENT_TYPE:
+            upload_metadata["ContentType"] = CONTENT_TYPE[extension]
 
         client.upload_fileobj(
             file,
             self.bucket_name,
-            str(pathlib.Path(self.folder_name, file_key)),
+            self._prepend_folder_name(file_key),
+            ExtraArgs=upload_metadata,
         )
 
         return file_key
@@ -161,13 +218,21 @@ class S3MediaStorage(MediaStorage):
         """
         A sync wrapper around :meth:`generate_file_url`.
         """
-        s3_client = self.get_client()
+        if self.sign_urls:
+            config = None
+        else:
+            from botocore import UNSIGNED
+            from botocore.config import Config
+
+            config = Config(signature_version=UNSIGNED)
+
+        s3_client = self.get_client(config=config)
 
         return s3_client.generate_presigned_url(
             ClientMethod="get_object",
             Params={
                 "Bucket": self.bucket_name,
-                "Key": str(pathlib.Path(self.folder_name, file_key)),
+                "Key": self._prepend_folder_name(file_key),
             },
             ExpiresIn=self.signed_url_expiry,
         )
@@ -191,7 +256,7 @@ class S3MediaStorage(MediaStorage):
         s3_client = self.get_client()
         response = s3_client.get_object(
             Bucket=self.bucket_name,
-            Key=str(pathlib.Path(self.folder_name, file_key)),
+            Key=self._prepend_folder_name(file_key),
         )
         return response["Body"]
 
@@ -215,7 +280,7 @@ class S3MediaStorage(MediaStorage):
         s3_client = self.get_client()
         return s3_client.delete_object(
             Bucket=self.bucket_name,
-            Key=str(pathlib.Path(self.folder_name, file_key)),
+            Key=self._prepend_folder_name(file_key),
         )
 
     async def bulk_delete_files(self, file_keys: t.List[str]):
@@ -231,7 +296,6 @@ class S3MediaStorage(MediaStorage):
 
         batch_size = 100
         iteration = 0
-        folder_name = self.folder_name
 
         while True:
             batch = file_keys[
@@ -247,7 +311,9 @@ class S3MediaStorage(MediaStorage):
                 Bucket=self.bucket_name,
                 Delete={
                     "Objects": [
-                        {"Key": str(pathlib.Path(folder_name, file_key))}
+                        {
+                            "Key": self._prepend_folder_name(file_key),
+                        }
                         for file_key in file_keys
                     ],
                 },
@@ -265,13 +331,16 @@ class S3MediaStorage(MediaStorage):
         start_after = None
 
         while True:
-            extra_kwargs: t.Dict[str, t.Any] = (
-                {"StartAfter": start_after} if start_after else {}
-            )
+            extra_kwargs: t.Dict[str, t.Any] = {}
+
+            if start_after:
+                extra_kwargs["StartAfter"] = start_after
+
+            if self.folder_name:
+                extra_kwargs["Prefix"] = f"{self.folder_name}/"
 
             response = s3_client.list_objects_v2(
                 Bucket=self.bucket_name,
-                Prefix=self.folder_name,
                 **extra_kwargs,
             )
 
@@ -286,9 +355,11 @@ class S3MediaStorage(MediaStorage):
                 # https://github.com/nedbat/coveragepy/issues/772
                 break  # pragma: no cover
 
-        prefix = f"{self.folder_name}/"
-
-        return [i.lstrip(prefix) for i in keys]
+        if self.folder_name:
+            prefix = f"{self.folder_name}/"
+            return [i.lstrip(prefix) for i in keys]
+        else:
+            return keys
 
     async def get_file_keys(self) -> t.List[str]:
         """